compiler.h: Change pointer integer casting to uintptr_t
Change pointer integer casting from unsigned long to uintptr_t
to improve the code portibility on the architecture which the
pointer is not an unsigned long integer (e.g. CHERI).

+++ include/linux/compiler.h
@@ -303,7 +303,7 @@ unsigned long read_word_at_a_time(const void *addr)
  */
 static inline void *offset_to_ptr(const int *off)
 {
-	return (void *)((unsigned long)off + *off);
+	return (void *)((uintptr_t)off + *off);
 }
 
 #endif /* __ASSEMBLY__ */
